Spotlight::BuildNGQuad: uBaseN is an int, not an unsigned short (90.04979 -> 90.622406)
The image forms n - 1 as a 32-bit loop invariant (subi r4, r5, 0x1 at
8282D114) and adds it to base (add r31, r4, r11 at 8282D134).  With
uBaseN declared unsigned short, MSVC demoted the whole expression to
16-bit arithmetic and hoisted n + 0xffff instead.  uPrev stays
unsigned short: the image's base + 0xffff (add r30, r11, r6 at
8282D13C) is exactly that demotion.

+++ src/system/world/Spotlight.cpp
@@ -1790,7 +1790,7 @@ void Spotlight::BuildNGQuad(BeamDef &def, RndTransformable::Constraint constrain
     for (int row = 0; row < nMinus1; row++) {
         for (int col = 0; col < nMinus1; col++) {
             int base = row + 1 + col * n;
-            unsigned short uBaseN = base + n - 1;
+            int uBaseN = base + n - 1;
             unsigned short uPrev = base - 1;
             if (iFace & 2) {
                 faces[iFace++].Set(uBaseN, uPrev, base + n);
